+++ SpringMvc/Models/POCO/OrderEntry.cs
@@ -11,6 +11,6 @@ namespace SpringMvc.Models.POCO
         public virtual BookType BookType { get; set; }
         public virtual Order Order { get; set; }
         public virtual int Amount { get; set; }
-        public virtual double Price { get; set; }
+        public virtual Decimal Price { get; set; }
     }
 }
